add it package for integration tests
