+++ spring-cloud-alibaba-starters/spring-cloud-alibaba-kubernetes/spring-cloud-starter-alibaba-kubernetes-config/src/test/resources/application-configmap.yml
@@ -7,7 +7,7 @@ spring:
           - name: my-configmap
             refreshable: true
   application:
-    name: normal
+    name: configmap
 logging:
   level:
     com.alibaba.kubernetes.config: debug
