+++ spring-cloud-alibaba-starters/spring-cloud-alibaba-kubernetes/spring-cloud-starter-alibaba-kubernetes-config/src/test/resources/configmap/configmap.yaml
